only warn about non-existent database if nix-locate fails
When using a non-standard database location for nix-index (e.g. via the
NixOS modules in the github:Mic92/nix-index-database flake), comma warns
that the database is non-existent even though it does exist and
nix-locate works fine.

This changes the logic so that it only checks if the database exists
when the `nix-locate` command fails.