Replace wrapProgram with substituteInPlace so PATH isn't polluted
Previously, when a command was executed with `, blah`,
it would inherit the $PATH var from comma. This meant
that there was some slight pollution, primarily since
it would be the `nix`, `fzy`, and `nix-index` bin paths
on PATH. Specifically they would be prefixed, so they would
override any other entries. This could lead to confusing
behavior if comma is built from a flake without overriding
the nixpkgs input.

This changes it so that the paths are directly embedded
in the binary, so no PATH pollution occurs.